# services/tours-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing tours.",
        "title": "Tours Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api",
    "paths": {
        "/tours": {
            "get": {
                "description": "Vraća listu svih tura dostupnih u sistemu.",
                "produces": [
                    "application/json"
                ],
                "summary": "Prikaz svih tura",
                "responses": {
                    "200": {
                        "description": "Lista svih tura",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Tour"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Kreira novu turu sa statusom 'draft' i cenom 0. Zahteva autentikaciju.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Kreiranje nove ture",
                "parameters": [
                    {
                        "description": "Podaci o turi (bez statusa, cene i authorId)",
                        "name": "tour",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Uspešno kreirana tura",
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    },
                    "400": {
                        "description": "Greška: Neispravan format zahteva",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Greška: Korisnik nije autorizovan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tours/my-tours": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Vraća listu svih tura koje je kreirao autor čiji se token koristi.",
                "produces": [
                    "application/json"
                ],
                "summary": "Prikaz tura kreiranih od strane ulogovanog autora",
                "responses": {
                    "200": {
                        "description": "Lista tura",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Tour"
                            }
                        }
                    },
                    "401": {
                        "description": "Greška: Korisnik nije autorizovan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/tours/{id}": {
            "get": {
                "description": "Vraća detalje specifične ture na osnovu njenog ID-ja.",
                "produces": [
                    "application/json"
                ],
                "summary": "Prikaz jedne ture po ID-ju",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji ture",
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    },
                    "404": {
                        "description": "Greška: Tura nije pronađena",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}/archive": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menja status ture u 'archived'.",
                "produces": [
                    "application/json"
                ],
                "summary": "Arhiviranje ture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno arhivirana tura",
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    },
                    "400": {
                        "description": "Greška: Tura nije objavljena",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Greška: Korisnik nije autorizovan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Greška: Nemate dozvolu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}/keypoints": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Dodaje novu ključnu tačku na turu sa datim ID-jem. Samo autor ture može dodati tačku.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Dodavanje ključne tačke na turu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Podaci o ključnoj tački",
                        "name": "keyPoint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TourKeyPoint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno dodata ključna tačka",
                        "schema": {
                            "$ref": "#/definitions/domain.TourKeyPoint"
                        }
                    }
                }
            }
        },
        "/tours/{id}/keypoints/{keypointId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ažurira postojeću ključnu tačku na turi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Ažuriranje ključne tačke",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID Ključne tačke",
                        "name": "keypointId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novi podaci o ključnoj tački",
                        "name": "keyPoint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TourKeyPoint"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirana ključna tačka",
                        "schema": {
                            "$ref": "#/definitions/domain.TourKeyPoint"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Briše ključnu tačku sa ture.",
                "produces": [
                    "application/json"
                ],
                "summary": "Brisanje ključne tačke",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID Ključne tačke",
                        "name": "keypointId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Poruka o uspehu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}/publish": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menja status ture u 'published' ako su ispunjeni svi uslovi.",
                "produces": [
                    "application/json"
                ],
                "summary": "Objavljivanje ture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno objavljena tura",
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    },
                    "400": {
                        "description": "Greška: Uslovi za objavljivanje nisu ispunjeni",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Greška: Korisnik nije autorizovan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Greška: Nemate dozvolu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}/reactivate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menja status arhivirane ture nazad u 'published'.",
                "produces": [
                    "application/json"
                ],
                "summary": "Ponovno aktiviranje ture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno reaktivirana tura",
                        "schema": {
                            "$ref": "#/definitions/domain.Tour"
                        }
                    },
                    "400": {
                        "description": "Greška: Tura nije arhivirana",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Greška: Korisnik nije autorizovan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Greška: Nemate dozvolu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}/reviews": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Dodaje novu recenziju na turu sa datim ID-jem. Zahteva autentikaciju.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Dodavanje recenzije na turu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Ture",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Podaci o recenziji",
                        "name": "review",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TourReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Poruka o uspehu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Tour": {
            "type": "object",
            "required": [
                "description",
                "difficulty",
                "name"
            ],
            "properties": {
                "archivedAt": {
                    "description": "Vreme arhiviranja",
                    "type": "string"
                },
                "authorId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "type": "integer"
                },
                "distance": {
                    "description": "--- NOVA POLJA ---",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "keyPoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TourKeyPoint"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "publishedAt": {
                    "description": "Vreme objave",
                    "type": "string"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TourReview"
                    }
                },
                "status": {
                    "description": "IZMENA: Koristimo novi tip",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.TourStatus"
                        }
                    ]
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "transportInfo": {
                    "description": "Lista vremena putovanja",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TourTransport"
                    }
                }
            }
        },
...
        "domain.TourStatus": {
            "type": "string",
            "enum": [
                "draft",
                "published",
                "archived"
            ],
            "x-enum-varnames": [
                "TourStatusDraft",
                "TourStatusPublished",
                "TourStatusArchived"
            ]
        },
        "domain.TourTransport": {
            "type": "object",
            "properties": {
                "timeInMinutes": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/domain.TransportType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}